+++ CHANGELOG.md
@@ -16,7 +16,7 @@ Go to the `v1` branch to see the changelog of Lume 1.
 ### Fixed
 - Issues running `site.build()` multiple times in the same process.
 - `mergedKeys` key is not merged.
-- Updated dependencies: `bar`, `unocss`, `tailwindcss`, `esbuild`, `std`, `deno/loader`, `sass`, `xml`, `magic-string` and some plugins.
+- Updated dependencies: `bar`, `unocss`, `tailwindcss`, `esbuild`, `std`, `deno/loader`, `sass`, `xml`, `magic-string`, `satori`, `mdx` and some icons.
 
 ## [3.0.6] - 2025-08-07
 ### Added