+++ CHANGELOG.md
@@ -6,10 +6,10 @@ and this project try to adheres to [Semantic Versioning](https://semver.org/).
 Go to the `v2` branch to see the changelog of Lume 2.
 Go to the `v1` branch to see the changelog of Lume 1.
 
-## [3.0.7] - Unreleased
+## [3.0.7] - 2025-09-02
 ### Added
 - `check_urls` plugin: New option `anchors` to check the URL hash.
-- `filter_pages` plugin: New option `beforeRender` to filter the pages before is rendered.
+- `filter_pages` plugin: New option `beforeRender` to filter the pages before rendering.
 - `inline` plugin: New option `sourceURL` to add the `sourceURL=inline:...` pragma to inlined JS and CSS files [#779].
 - `site.initDebugBar()` to initialize manually the `debugBar`.
 - `gh:` specifier for `add()` and `copy()`, to use files from GitHub repositories.
@@ -218,7 +218,7 @@ Go to the `v1` branch to see the changelog of Lume 1.
 [#779]: https://github.com/lumeland/lume/issues/779
 [#782]: https://github.com/lumeland/lume/issues/782
 
-[3.0.7]: https://github.com/lumeland/lume/compare/v3.0.6...HEAD
+[3.0.7]: https://github.com/lumeland/lume/compare/v3.0.6...v3.0.7
 [3.0.6]: https://github.com/lumeland/lume/compare/v3.0.5...v3.0.6
 [3.0.5]: https://github.com/lumeland/lume/compare/v3.0.4...v3.0.5
 [3.0.4]: https://github.com/lumeland/lume/compare/v3.0.3...v3.0.4